+++ R/PersonClass.R
@@ -1,6 +1,7 @@
 Person <- setRefClass("Person", 
                       
-                      fields = list(raw_responses = 'integer',
+                      fields = list(raw_responses = 'character',
+                                    raw_responses_location = 'integer',
                                     responses = 'integer',
                                     items_answered = 'integer',
                                     thetas = 'matrix',
@@ -16,7 +17,8 @@ Person <- setRefClass("Person",
                          initialize = function(nfact, nitems, thetas.start_in, score,
                                                theta_SEs){
                              'Initialize the person object given background information'
-                             raw_responses <<- as.integer(rep(NA, nitems))
+                             raw_responses <<- as.character(rep(NA, nitems))
+                             raw_responses_location <<- as.integer(rep(NA, nitems))
                              responses <<- as.integer(rep(NA, nitems))
                              valid_item <<- rep(TRUE, nitems)
                              items_answered <<- as.integer(rep(NA, nitems))